Fix some issues with byte counts

+++ include/RaspiLowLevel.hpp
@@ -24,12 +24,12 @@ namespace RASPI {
 	private:
 		bool init;
 		bool spi_init;
-		int stm32_init_string[11];
-		int stm32_receive_string[11];
-		int stm32_pair_string[11];
-		int stm32_accept_string[11];
-		int raw_spi_data[21];
-		int dummy_string[21];
+		uint8_t stm32_init_string[11];
+		uint8_t stm32_receive_string[11];
+		uint8_t stm32_pair_string[11];
+		uint8_t stm32_accept_string[11];
+		uint8_t raw_spi_data[21];
+		uint8_t dummy_string[21];
 		std::vector<uint16_t> buffer;
 	};
